+++ Source/Scene/GlobeSurfaceTileProvider.js
@@ -127,6 +127,7 @@ define([
         this.oceanNormalMap = undefined;
         this.zoomedOutOceanSpecularIntensity = 0.5;
         this.enableLighting = false;
+        this.showGroundAtmosphere = false;
         this.shadows = ShadowMode.RECEIVE_ONLY;
 
         this._quadtree = undefined;
@@ -877,6 +878,9 @@ define([
             u_lightingFadeDistance : function() {
                 return this.properties.lightingFadeDistance;
             },
+            u_nightFadeDistance : function() {
+                return this.properties.nightFadeDistance;
+            },
             u_center3D : function() {
                 return this.properties.center3D;
             },
@@ -982,6 +986,7 @@ define([
                 zoomedOutOceanSpecularIntensity : 0.5,
                 oceanNormalMap : undefined,
                 lightingFadeDistance : new Cartesian2(6500000.0, 9000000.0),
+                nightFadeDistance : new Cartesian2(10000000.0, 40000000.0),
 
                 center3D : undefined,
                 rtc : new Cartesian3(),
@@ -1135,6 +1140,29 @@ define([
     })();
 
     var otherPassesInitialColor = new Cartesian4(0.0, 0.0, 0.0, 0.0);
+    var surfaceShaderSetOptionsScratch = {
+        frameState : undefined,
+        surfaceTile : undefined,
+        numberOfDayTextures : undefined,
+        applyBrightness : undefined,
+        applyContrast : undefined,
+        applyHue : undefined,
+        applySaturation : undefined,
+        applyGamma : undefined,
+        applyAlpha : undefined,
+        applySplit : undefined,
+        showReflectiveOcean : undefined,
+        showOceanWaves : undefined,
+        enableLighting : undefined,
+        showGroundAtmosphere : undefined,
+        perFragmentGroundAtmosphere : undefined,
+        hasVertexNormals : undefined,
+        useWebMercatorProjection : undefined,
+        enableFog : undefined,
+        enableClippingPlanes : undefined,
+        clippingPlanes : undefined,
+        clippedByBoundaries : undefined
+    };
 
     function addDrawCommandsForTile(tileProvider, tile, frameState) {
         var surfaceTile = tile.data;
@@ -1157,8 +1185,26 @@ define([
         var showOceanWaves = showReflectiveOcean && defined(oceanNormalMap);
         var hasVertexNormals = tileProvider.terrainProvider.ready && tileProvider.terrainProvider.hasVertexNormals;
         var enableFog = frameState.fog.enabled;
+        var showGroundAtmosphere = tileProvider.showGroundAtmosphere;
         var castShadows = ShadowMode.castShadows(tileProvider.shadows);
         var receiveShadows = ShadowMode.receiveShadows(tileProvider.shadows);
+
+        var perFragmentGroundAtmosphere = false;
+        if (showGroundAtmosphere) {
+            var mode = frameState.mode;
+            var camera = frameState.camera;
+            var cameraDistance;
+            if (mode === SceneMode.SCENE2D || mode === SceneMode.COLUMBUS_VIEW) {
+                cameraDistance = camera.positionCartographic.height;
+            } else {
+                cameraDistance = Cartesian3.magnitude(camera.positionWC);
+            }
+            var fadeOutDistance = tileProvider.nightFadeOutDistance;
+            if (mode !== SceneMode.SCENE3D) {
+                fadeOutDistance -= frameState.mapProjection.ellipsoid.maximumRadius;
+            }
+            perFragmentGroundAtmosphere = cameraDistance > fadeOutDistance;
+        }
 
         if (showReflectiveOcean) {
             --maxTextures;
@@ -1229,6 +1275,18 @@ define([
             }
         }
 
+        var surfaceShaderSetOptions = surfaceShaderSetOptionsScratch;
+        surfaceShaderSetOptions.frameState = frameState;
+        surfaceShaderSetOptions.surfaceTile = surfaceTile;
+        surfaceShaderSetOptions.showReflectiveOcean = showReflectiveOcean;
+        surfaceShaderSetOptions.showOceanWaves = showOceanWaves;
+        surfaceShaderSetOptions.enableLighting = tileProvider.enableLighting;
+        surfaceShaderSetOptions.showGroundAtmosphere = showGroundAtmosphere;
+        surfaceShaderSetOptions.perFragmentGroundAtmosphere = perFragmentGroundAtmosphere;
+        surfaceShaderSetOptions.hasVertexNormals = hasVertexNormals;
+        surfaceShaderSetOptions.useWebMercatorProjection = useWebMercatorProjection;
+        surfaceShaderSetOptions.clippedByBoundaries = surfaceTile.clippedByBoundaries;
+
         var tileImageryCollection = surfaceTile.imagery;
         var imageryIndex = 0;
         var imageryLen = tileImageryCollection.length;
@@ -1287,6 +1345,8 @@ define([
             uniformMapProperties.oceanNormalMap = oceanNormalMap;
             uniformMapProperties.lightingFadeDistance.x = tileProvider.lightingFadeOutDistance;
             uniformMapProperties.lightingFadeDistance.y = tileProvider.lightingFadeInDistance;
+            uniformMapProperties.nightFadeDistance.x = tileProvider.nightFadeOutDistance;
+            uniformMapProperties.nightFadeDistance.y = tileProvider.nightFadeInDistance;
             uniformMapProperties.zoomedOutOceanSpecularIntensity = tileProvider.zoomedOutOceanSpecularIntensity;
 
             uniformMapProperties.center3D = surfaceTile.center;
@@ -1414,7 +1474,19 @@ define([
                 uniformMap = combine(uniformMap, tileProvider.uniformMap);
             }
 
-            command.shaderProgram = tileProvider._surfaceShaderSet.getShaderProgram(frameState, surfaceTile, numberOfDayTextures, applyBrightness, applyContrast, applyHue, applySaturation, applyGamma, applyAlpha, applySplit, showReflectiveOcean, showOceanWaves, tileProvider.enableLighting, hasVertexNormals, useWebMercatorProjection, applyFog, clippingPlanesEnabled, clippingPlanes, surfaceTile.clippedByBoundaries);
+            surfaceShaderSetOptions.numberOfDayTextures = numberOfDayTextures;
+            surfaceShaderSetOptions.applyBrightness = applyBrightness;
+            surfaceShaderSetOptions.applyContrast = applyContrast;
+            surfaceShaderSetOptions.applyHue = applyHue;
+            surfaceShaderSetOptions.applySaturation = applySaturation;
+            surfaceShaderSetOptions.applyGamma = applyGamma;
+            surfaceShaderSetOptions.applyAlpha = applyAlpha;
+            surfaceShaderSetOptions.applySplit = applySplit;
+            surfaceShaderSetOptions.enableFog = applyFog;
+            surfaceShaderSetOptions.enableClippingPlanes = clippingPlanesEnabled;
+            surfaceShaderSetOptions.clippingPlanes = clippingPlanes;
+
+            command.shaderProgram = tileProvider._surfaceShaderSet.getShaderProgram(surfaceShaderSetOptions);
             command.castShadows = castShadows;
             command.receiveShadows = receiveShadows;
             command.renderState = renderState;